// src/types/fabric.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Fabric = { id: number, name: string | null, length: number | null, width: number | null, costs: number | null, fotoPath: string | null, producer: string | null, kindOfFabricId: number | null, dateOfPurchase: string | null, };